Extract webui-press bundler module
